backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.0.3

# AI/ML packages
llama-index==0.9.48
llama-index-llms-openai==0.1.15
llama-index-embeddings-openai==0.1.9
llama-index-llms-huggingface==0.2.0
llama-index-embeddings-huggingface==0.2.0
openai==1.12.0

# ML packages
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector database and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Document processing
unstructured>=0.10.0
pypdf>=3.15.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async and caching
redis>=5.0.0
celery>=5.3.0
httpx>=0.25.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-magic>=0.4.0